requirements-test.txt
# Python test dependencies for Incus Salt formula
#
# Install with: pip install -r requirements-test.txt
# Testing frameworks (required)
pytest>=7.4.0
pytest-xdist>=3.3.0              # Parallel test execution
pytest-cov>=4.1.0                 # Coverage reporting
pytest-timeout>=2.1.0             # Test timeouts
pytest-mock>=3.11.0               # Mocking support
assertpy

# Data handling (required)
PyYAML>=6.0                       # YAML parsing
requests>=2.31.0                  # HTTP requests

# Code quality (optional)
# pylint>=2.17.0                  # Python linting
# black>=23.7.0                   # Code formatting
# flake8>=6.0.0                   # Style checking
# mypy>=1.4.0                     # Type checking

# Documentation (optional)
# pytest-html>=3.2.0              # HTML test reports
# pytest-json-report>=1.5.0       # JSON test reports

# Development tools (optional)
# pytest-watch>=4.2.0             # Watch mode for tests
# ipython>=8.14.0                 # Interactive debugging
# ipdb>=0.13.13                   # Debugger
